Fix calculating ssim value.
data_range is required, due to the fact that library infers it on its own. Documentation says "By default, this is estimated from the image data-type.". I have tested it on this paper results: http://www.vision.ee.ethz.ch/~timofter/publications/Agustsson-CVPRW-2017suppl.pdf  . Without data_range it returns incorrect values, because of wrong range inference. With this fix, it seems to be working.